Pull sort logic into util function so it can be tested easily.
